Fixed symlink problems and batch analysis

+++ data_transfer_and_run.py
@@ -20,12 +20,18 @@ def transfer_data():
     for vid, result in zip(json_config["videos_to_anly"], json_config["server_dest_folder"]):
         vid_name = vid.split("/")[-1][:-4]
         origin = os.path.join(server_path, vid)
+        print(f"Transferring {vid} to {video_folder}")
         dest_folder = os.path.join(result_folder, vid_name)
         if not os.path.exists(dest_folder):
             os.makedirs(dest_folder, exist_ok=True)
 
-        shutil.copy(origin, video_folder)
-        subprocess.run(["sbatch", f"/home/{user}/LSENS-DeepLabCut/analyze_video.sbatch", f"{json_config['config_path']}", os.path.join(video_folder, vid_name, ".avi"), f"{dest_folder}"])
+        if not os.path.exists(os.path.join(video_folder, vid.split("/")[-1])):
+            shutil.copy(origin, video_folder)
+
+        command = f" {json_config['config_path']} {os.path.join(video_folder, vid_name + '.avi')} {dest_folder}"
+        print(f"Executing command: {command}")
+        subprocess.run(["echo", f"INFO: Launching dlc for video {vid_name}"])
+        os.system(f"sbatch /home/{user}/LSENS-DeepLabCut/launch_dlc.sbatch" + command)
 
 if __name__ == "__main__":
     transfer_data()
